requirements.txt
Flask
requests
gunicorn
python-dotenv

